refactor: store 부분 패키지 변경
-- 기존 도메인에 있던 부분 상위 application, presentation 으로 위치 변경
