+++ src/main/java/com/jangburich/domain/point/domain/repository/PointTransactionRepository.java
@@ -2,7 +2,7 @@ package com.jangburich.domain.point.domain.repository;
 
 import com.jangburich.domain.point.domain.PointTransaction;
 import com.jangburich.domain.entity.Store;
-import com.jangburich.domain.store.presentation.dto.response.store.StoreChargeHistoryResponse;
+import com.jangburich.presentation.store.dtos.response.store.StoreChargeHistoryResponse;
 import com.jangburich.domain.user.domain.User;
 import java.util.List;
