+++ src/main/java/com/jangburich/domain/repository/queryDsl/StoreQueryDslRepositoryImpl.java
@@ -1,8 +1,8 @@
 package com.jangburich.domain.repository.queryDsl;
 
 import com.jangburich.domain.entity.Category;
-import com.jangburich.domain.store.presentation.dto.response.store.QSearchStoresResponse;
-import com.jangburich.domain.store.presentation.dto.response.store.SearchStoresResponse;
+import com.jangburich.presentation.store.dtos.response.store.QSearchStoresResponse;
+import com.jangburich.presentation.store.dtos.response.store.SearchStoresResponse;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.jpa.impl.JPAQuery;
