--- src/main/java/com/jangburich/domain/store/presentation/dto/condition/StoreSearchCondition.java
@@ -1,7 +1,0 @@
-package com.jangburich.domain.store.presentation.dto.condition;
-
-public record StoreSearchCondition(
-        Double lat,
-        Double lon
-) {
-}
