--- src/main/java/com/jangburich/domain/store/presentation/dto/request/StoreAdditionalInfoCreateRequest.java
@@ -1,13 +1,0 @@
-package com.jangburich.domain.store.presentation.dto.request;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class StoreAdditionalInfoCreateRequest {
-	private Boolean reservationAvailable;
-	private Long maxReservation;
-	private Long minPrepayment;
-	private Long prepaymentDuration;
-}
