--- src/main/java/com/jangburich/domain/store/presentation/dto/response/store/StoreCreateResponseDto.java
@@ -1,7 +1,0 @@
-package com.jangburich.domain.store.presentation.dto.response.store;
-
-public record StoreCreateResponseDto (
-    String storeIssueCode,
-    String message
-) {
-}
