+++ src/main/java/com/jangburich/presentation/store/dtos/condition/StoreSearchCondition.java
@@ -1,0 +1,7 @@
+package com.jangburich.presentation.store.dtos.condition;
+
+public record StoreSearchCondition(
+        Double lat,
+        Double lon
+) {
+}
