+++ src/main/java/com/jangburich/presentation/store/dtos/response/store/StoreCreateResponseDto.java
@@ -1,0 +1,7 @@
+package com.jangburich.presentation.store.dtos.response.store;
+
+public record StoreCreateResponseDto (
+    String storeIssueCode,
+    String message
+) {
+}
