+++ src/main/java/com/jangburich/presentation/store/dtos/response/store/StoreMenu.java
@@ -1,0 +1,24 @@
+package com.jangburich.presentation.store.dtos.response.store;
+
+import com.querydsl.core.annotations.QueryProjection;
+
+public record StoreMenu(
+        Long menuId,
+        String menuName,
+        Boolean isSignatureMenu,
+        String description,
+        int price,
+        String menuImgUrl
+) {
+
+    @QueryProjection
+    public StoreMenu(Long menuId, String menuName, Boolean isSignatureMenu, String description, int price,
+                     String menuImgUrl) {
+        this.menuId = menuId;
+        this.menuName = menuName;
+        this.isSignatureMenu = isSignatureMenu;
+        this.description = description;
+        this.price = price;
+        this.menuImgUrl = menuImgUrl;
+    }
+}
