+++ src/test/java/com/jangburich/domain/store/service/provider/RandomNumberProviderTest.java
@@ -7,6 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 
+import com.jangburich.application.store.provider.RandomNumberProvider;
+
 @ActiveProfiles("local")
 @SpringBootTest
 class RandomNumberProviderTest {
